pyproject.toml
[tool.poetry]
name = "My-shop-bot"
version = "0.1.0"
description = "A project to develop a Telegram bot for tracking Wildberries orders"
authors = [
    "Nadezhda_Popova <nadezhdapopova13@yandex.ru>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
flask = ">=3.1.2,<4.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
flask-sqlalchemy = ">=3.1.1,<4.0.0"
alembic = ">=1.16.5,<2.0.0"
flask-migrate = ">=4.1.0,<5.0.0"
requests = ">=2.32.5,<3.0.0"
python-telegram-bot = {extras = ["webhooks"], version = "^22.5"}
apscheduler = ">=3.11.0,<4.0.0"


[build-system]
requires = "poetry-core>=2.0.0,<3.0.0"
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.18.2"
isort = "^6.0.1"
black = "^25.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[tool.mypy]
disallow_untyped_defs=true
warn_return_any=true
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.black]
line-length = 119
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.isort]
line_length = 119
multi_line_output = 3